==> 00 - Build transliteration tables.ipynb <==
## Build transliteration table

Pull table from riptsource.org/cms/scripts/page.php?item_id=entry_detail&uid=vsytndbyev and build both forward and reserve transliteration tables.

import requests
import pandas as pd
import unidecode
import numpy as np

user_agent = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/35.0.1916.47 Safari/537.36'
headers = {'User-Agent': user_agent}
response = requests.get('https://scriptsource.org/cms/scripts/page.php?item_id=entry_detail&uid=vsytndbyev', headers=headers)

results = pd.read_html(response.text)
df = results[1]
df.head()

We can't see the glyph because it was an image! So we'll convert the Unicode code - `USV` - to the character.

df.Glyph = df.USV.apply(lambda val: chr(int(val, 16)))
df.head()

We'll save this one for now as a CSV (although we probably won't use it)

df.to_csv("transliteration-table.csv", index=False)

## Simplified transliteration table

df = pd.read_csv("transliteration-table.csv")
df = df.drop(columns=['USV'])
df = df.set_index('Glyph')
df

We're really just looking at ASCII Latin transliteraitons, so we'll turn `hā` into `ha`, etc.

def simplify(value):
    try:
        return unidecode.unidecode(value).lower()
    except:
        return np.nan
    
simplified = df.applymap(simplify)
simplified

# Break apart h/ha into h and ha
uniqued = simplified.fillna('').apply(lambda row: list(set([v for v in row.values if v and '/' not in str(v)])), axis=1)

# Remove anything that involves @ or ~
uniqued = uniqued.apply(lambda row: [v for v in row if '@' not in v])

# Remove ~ (but keep the rest)
uniqued = uniqued.apply(lambda row: [v.replace('~', '') for v in row])

uniqued

### Add punctuation

punctuation = {
    '።': ['.'],
    '፡': [' '],
    '፣': [','],
    '፤': [';'],
    '፥': [':'],
    '፧': ['?'],
    '፦': [':']
}

uniqued = pd.concat([pd.Series(punctuation), uniqued])
uniqued

### Save

uniqued.to_json("eth_lat.json")

import json

with open("eth_lat.json") as f:
    eth_lat = json.load(f)

eth_lat

lat_eth = {}
for eth in eth_lat.keys():
    for lat in eth_lat[eth]:
        if lat not in lat_eth.keys():
            lat_eth[lat] = []
        lat_eth[lat].append(eth)

for key in lat_eth.keys():
    lat_eth[key] = list(set(lat_eth[key]))

lat_eth

with open("lat_eth.json", 'w') as f:
    json.dump(lat_eth, f)
