==> 01 - Create randomized transliterations.ipynb <==
# Import lookup table

import random
import json
import string
import unidecode
import regex

with open("raw/lookup-table.json") as f:
    lookup = json.load(f)

lookup

# Tools to build randomized transliterations

import re

failures = set([])

def select_char(c):
    # If it's not in the lookup table, that's ok
    try:
        return random.choice(lookup[c])
    except:
        failures.add(c)
        return unidecode.unidecode(c).lower()

def transliterate(word):
    selected = ''.join([select_char(c) for c in word])
    selected = selected.replace("፡", " ")
    selected = re.sub(r'\s+', ' ', selected)
    return selected.strip()

text = "ማንም፡ሰው፡ቢሆን፡የጭካኔ፡ስቃይ፡እንዳይደርስበት፡ወይም፡ከሰብዓዊ፡አፈጻጸም፡ውጭ፡የሆነ፡የተዋረድ፡ተግባር፡ወይም፡ቅጣት፡አይፈጸምበትም።"

transliterate(text)

for _ in range(5):
    result = transliterate(text)
    print(f"* {result}")

# Use it

Using data from `new-am.txt`, which is Bible + legal + news.

#text = open("corpuscrawler-master/corpus/am.txt").readlines()
text = open("raw/new-am.txt").read().strip().splitlines()
text[:5]

# Drop duplicates
uniques = [line for line in list(set(text)) if line]
len(uniques)

# Transliterate and save

import pandas as pd

def clean(text):
    # Remove invalid unicode chars
    # note it's the regex library, not re
    cleaned = regex.sub(r'(\p{Cn}|\p{Co}|\p{Cf}|\p{Cc})', '', text)
    
    # Replace space char with spaces
    cleaned = text.replace("፡", " ").strip()

    # Remove duplicate spaces
    cleaned = re.sub(r'\s+', ' ', cleaned)
    return cleaned

# Do you want to transliterate each row in multiple ways?
NUM_TRANSLITERATIONS = 2

sources = []
transliterateds = []
skipped = 0
for text in uniques:
    cleaned = clean(text)
    options = [transliterate(cleaned) for _ in range(NUM_TRANSLITERATIONS)]
    for transliteration in options:
        # Make sure we can split afterwards, some of these texts are real weird
        if len(cleaned.split(" ")) != len(transliteration.split(" ")):
            skipped = skipped + 1
        else:
            sources.append(cleaned)
            transliterateds.append(transliteration)

print("Skipped", skipped)
print("Fail chars:", failures)

# Shuffle so it's randomized
df = pd.DataFrame({
    'source': sources,
    'transliterated': transliterateds
}).sample(frac=1).reset_index(drop=True)

print(df.shape)
df.head()

with open('raw/original.txt', 'w') as f:
    f.write('\n'.join(df.source))
    
with open('raw/transliterated.txt', 'w') as f:
    f.write('\n'.join(df.transliterated))

!head -n 2 raw/original.txt

!head -n 2 raw/transliterated.txt
