==> 02 - OpenNMT char-based.ipynb <==
# Creating test and train datasets

!unzip Archive.zip

import pandas as pd
import re

original = open("original.txt").read()
transliterated = open("transliterated.txt").read()

words = pd.DataFrame({
    'orig': re.split(r'\s+', original),
    'trans': re.split(r'\s+', transliterated)
})
words.head()

words.orig = words.orig.str.join(' ')
words.trans = words.trans.str.join(' ')

words.head(10)

from sklearn.model_selection import train_test_split

train, test = train_test_split(words, test_size=5000)

print("Training on", train.shape)
print("Testing on", test.shape)

!mkdir -p data

# Save as training
# Can't use .to_csv because of quotins
with open('data/src-train.txt', 'w') as f:
    f.write('\n'.join(train.trans))
with open('data/tgt-train.txt', 'w') as f:
    f.write('\n'.join(train.orig))

with open('data/src-val.txt', 'w') as f:
    f.write('\n'.join(test.trans))
with open('data/tgt-val.txt', 'w') as f:
    f.write('\n'.join(test.orig))

!head -n 5 data/tgt-val.txt

train.head()

# Data setup

Following quickstart instructions from https://github.com/OpenNMT/OpenNMT-py#quickstart.

## Training

I'm just using the terminal commands because the Python bindings were just Too Much Work.

!pip install OpenNMT-py

## Preprocess

!onmt_preprocess \
    -train_src data/src-train.txt \
    -train_tgt data/tgt-train.txt \
    -valid_src data/src-val.txt \
    -valid_tgt data/tgt-val.txt \
    -save_data data/demo \
    -overwrite

## Train

# Change to false to get GPU power on Colab
if False:
    !onmt_train \
        -data data/demo \
        -save_model demo-model \
        --valid_steps 50 \
        --train_steps 2 \
        --early_stopping 5
else:
    !CUDA_VISIBLE_DEVICES=0 \
        onmt_train \
        -world_size 1 \
        -gpu_ranks 0 \
        -data data/demo \
        -save_model demo-model \
        --valid_steps 3000 \
        --train_steps 50000 \
        --early_stopping 3

# Testing it out

Testing the text below, you can see the "answer" on Google translate [right here](https://translate.google.com/#view=home&op=translate&sl=am&tl=en&text=%E1%8A%A5%E1%8A%94%E1%88%9D%E1%8D%A5%20%E1%8B%88%E1%8A%95%E1%8B%B5%E1%88%9E%E1%89%BD%20%E1%88%86%E1%8B%AD%E1%8D%A5%20%E1%8B%A8%E1%88%A5%E1%8C%8B%20%E1%8A%A5%E1%8A%95%E1%8B%B0%20%E1%88%98%E1%88%86%E1%8A%93%E1%89%BD%E1%88%81%E1%8D%A5%0A%E1%89%A0%E1%8A%AD%E1%88%AD%E1%88%B5%E1%89%B6%E1%88%B5%E1%88%9D%20%E1%88%95%E1%8D%83%E1%8A%93%E1%89%B5%20%E1%8A%A5%E1%8A%95%E1%8B%B0%20%E1%88%98%E1%88%86%E1%8A%93%E1%89%BD%E1%88%81%20%E1%8A%A5%E1%8A%95%E1%8C%82%0A%E1%88%98%E1%8A%95%E1%8D%88%E1%88%B3%E1%8B%8D%E1%8B%AB%E1%8A%95%20%E1%8A%A5%E1%8A%95%E1%8B%B0%20%E1%88%98%E1%88%86%E1%8A%93%E1%89%BD%E1%88%81%20%E1%88%8D%E1%8A%93%E1%8C%88%E1%88%AB%E1%89%BD%E1%88%81%0A%E1%8A%A0%E1%88%8D%E1%89%BB%E1%88%8D%E1%88%81%E1%88%9D%E1%8D%A2).

!head -n 1 original.txt

!head -n 1 transliterated.txt

import subprocess

# Pull the first 10 originals and transliterateds
originals = subprocess.run("head -n 500 original.txt",
                            shell=True,
                            stdout=subprocess.PIPE).stdout.decode("utf-8").strip()
transliterated = subprocess.run("head -n 500 transliterated.txt",
                            shell=True,
                            stdout=subprocess.PIPE).stdout.decode("utf-8").strip()

# You can also use other stuff
originals = open("sera_am.txt").read()
transliterated = open("sera_rom.txt").read()

!pip install unidecode

import unidecode

def clean(lat_word):
  return unidecode.unidecode(lat_word).lower()

def add_spaces(word):
  return ' '.join(word)

orig_test = []
trans_test = []

# Go through each line, making sure they
# have the same number of words before you
# add them to the list
for o, t in zip(originals.splitlines(), transliterated.splitlines()):
  o_words = [add_spaces(w) for w in re.split(r"\s+", o) if w]
  # - is sometimes transliterated as a space
  t_words = [add_spaces(w) for w in re.split(r"\s+", clean(t)) if w]
  if(len(o_words) == len(t_words)):
    orig_test.extend(o_words)
    trans_test.extend(t_words)
  else:
    print("Not a matched pair:", o, t)

# import re

# orig_test = [' '.join(word) for word in re.split('\s+', originals)]
# trans_test = [' '.join(word) for word in re.split('\s+', transliterated)]
with open("data/test.txt", 'w') as f:
    f.write('\n'.join(trans_test))

!head -n 5 data/test.txt

print(len(trans_test), trans_test[:5])
print(len(orig_test), orig_test[:5])

**You'll need to change the model name in `onmt_translate` below.** It's probably the most recently changed model file, so at the top of this list: 

model_name = subprocess.run("ls -t *model* | head -n 1",
                            shell=True,
                            stdout=subprocess.PIPE).stdout.decode("utf-8").strip()
print("Using model", model_name)

!onmt_translate \
    -model {model_name} \
    -src data/test.txt \
    -output data/pred.txt -replace_unk


results = open("data/pred.txt").read().splitlines()
test_results = pd.DataFrame({
    'result': results,
    'original': orig_test,
    'transliterated': trans_test
})
print((test_results.result == test_results.original).value_counts(normalize=True))
test_results
